feat: Add rtt_target
And then use rtt_init_println, rprintln and debug_rprintln. If this is run
in dev profile both debug_rprintln and rprintln lines are printed if run
in release profile only rprintln is printed.

Also, adding only rtt_init_print and commenting out the two println! we
don't get any errors messagess from when running. But if we comment out
also rtt_init_print we get an error and `use rtt_target` we get an error:

 ERROR probe_rs::cmd::run: Failed to attach to RTT continuing...